requirements.txt
basic-pitch==0.2.6
pyboy==1.6.9
python-dotenv==1.0.0
pytube==15.0.0
librosa==0.9.1
pretty_midi==0.2.10
pydub==0.25.1
PyAudio==0.2.14
Pillow==10.1.0
pygame==2.5.2
piano-visualizer==1.1.0
pyFluidSynth==1.3.2
moviepy==1.0.3
midi2audio==0.1.1
